Update 1.md
Some minor tests.
It seems like speaking the story aloud is a wee bit faster than editing by hand, given that I can't find a prebuilt solution for dealing with the 'long s' problem.
But that will take quite some time.
Also if I do that, it seems that it has trouble with the name Circe. Perhaps, I should say something like "the enchantress" and then ctrl-h it. 

Finally, if it's an imperfect transcription anyways, then I may as well not sweat the details on getting everything exact or in full.